feat(CheckMode): Implement SeperatorError checking and fixing
SeperatorError needs multiple lines to check if there is an error. So
the abstract MarkdownError validate method had to be changed to accept a
list of strings. This does mean that every validation function should
check if it got the required amount of strings.

To be able to still supply MathError with one line the error checking
loop is split up into two loops. 1 for validators that take 1 line and
one for validators that take multiple lines.