feat(CheckMode): Implement watching for changes
Added -w / --watch flags (argparse)

Using watchdog (module) the input directory or file is watched for
changes. When a change occurs the modified/new file is converted.

The conversion is the same as the regular conversion process, the output
path is also created with subdirectories etc...

Disabled the TqdmLogger because it is broken, and throws errors.