feat(CheckMode): Disable the errors that are flagged to be disabled
Before adding the error the error to the `errors` list they are now
checked to see if they are disabled.

Disabled error names are stored as part of the class and therefore do
not have to be passed around.

Possible optimization would be to disable the error before even checking
for them. This would require a change in the way errors are currently
found (a.k.a hardcoded).

Still a WIP.