feat(ConvertMode): WIP - Show pandoc errors (#36)
The errors show up now. There is a difference detected between an
warning and an actual error. Warnings are printed in yellow and errors
in red.

Errors and warnings are always added to the log. When visual mode is
active, it is passed as an debug message (to prevent double printing).
When not in visual mode it is printed as an warning, since it should not
stop the whole program.

This commit does not handle the printing of errors/warnings when
converting a whole directory. The progress bar may not work correctly.
This will be addressed in the next commit(s).

This commit also breaks the tests because it changed the function
signature for the `subprocess.run` call. These tests will be fixed and
new ones will be added.