fix: Print the warnings and errors correctly with the progress bar.
When converting a directory `self._converting_dir` is set to true. This
is checked when printing the warnings and error messages. If it is set
to true it uses the tqdm writer function to print the messages. This way
it gets printed above the progress bar which keeps the progress bar
intact.