feat(Config): Add support for config files
Totally re-write how arguments are parsed to allow for config files to
be parsed.

The config class now keeps track of the 'options' both the config file
parser and the arguments parser write to these options. However the
argument parser overwrites the values from the config parser (because
these arguments are parsed later). This creates the behaviour that the
command line arguments overwrite the config file.

Note: this commit will break some tests, because it changes function
signatures/passed arguments. It is also not tested at all.

Ref: #35