fix(config): Adding required arguments only when flags
The required argument can only be passed to the argument parser when
adding a flag. Positional arguments are always required and argparse
raises an error when adding the required argument.

The logic to check if something was incorrect. It did not add required to
Positional arguments because of the first check but it also did not add
it to flags. This is fixed.