fix: Remove progress bar when in verbose (INFO) mode.
When verbose mode is enabled the progress bar is hidden. This is done
because the logger does not use tqdm.write which means it messes up the
progress bar. Next to that, there is so many output that it becomes not
readably anyway because it is meant for log files, which do not need a
progress bar anyway.

Closes: #23